adding content

+++ app/Models/Project.php
@@ -33,9 +33,11 @@ class Project extends Model
     public function setImageAttribute($image)
     {
         if ($image) {
-            $oldFile = public_path('upload/images/' . $this->attributes['image']);
-            if (File::exists($oldFile)) {
-                File::delete($oldFile);
+            if (isset($this->attributes['image'])) {
+                $oldFile = public_path('upload/images/' . $this->attributes['image'] ?? '');
+                if (File::exists($oldFile)) {
+                    File::delete($oldFile);
+                }
             }
             $name = time() . '_' . $image->getClientOriginalName();
             $image->move('upload/images/', $name);
